Move berry pots app rodata to C
