+++ src/alph_puzzle.c
@@ -434,17 +434,95 @@ static const GraphicsModes sGraphicsMode = {
     GX_BG0_AS_2D,
 };
 
-static const BgTemplate sBgTemplate5 = {0, 0, GF_BG_BUF_SIZE_256x256_4BPP, 0x0000, GF_BG_SCR_SIZE_256x256, GX_BG_COLORMODE_16, 30, 0, 0, 2, 0, 0, 0};
+static const BgTemplate sBgTemplate5 = {
+    0, 0,
+    GF_BG_BUF_SIZE_256x256_4BPP,
+    0x000,
+    GF_BG_SCR_SIZE_256x256,
+    GX_BG_COLORMODE_16,
+    30,
+    0,
+    0,
+    2,
+    0,
+    0,
+    FALSE
+};
 
-static const BgTemplate sBgTemplate1 = {0, 0, GF_BG_BUF_SIZE_256x256_4BPP, 0x0000, GF_BG_SCR_SIZE_256x256, GX_BG_COLORMODE_16, 31, 4, 0, 1, 0, 0, 0};
+static const BgTemplate sBgTemplate1 = {
+    0, 0,
+    GF_BG_BUF_SIZE_256x256_4BPP,
+    0x000,
+    GF_BG_SCR_SIZE_256x256,
+    GX_BG_COLORMODE_16,
+    GX_BG_SCRBASE_0xf800,
+    GX_BG_CHARBASE_0x10000,
+    GX_BG_EXTPLTT_01,
+    1,
+    GX_BG_AREAOVER_XLU,
+    0,
+    FALSE
+};
 
-static const BgTemplate sBgTemplate6 = {0, 0, GF_BG_BUF_SIZE_256x256_4BPP, 0x0000, GF_BG_SCR_SIZE_256x256, GX_BG_COLORMODE_16, 29, 0, 0, 3, 0, 0, 0};
+static const BgTemplate sBgTemplate6 = {
+    0, 0,
+    GF_BG_BUF_SIZE_256x256_4BPP,
+    0x000,
+    GF_BG_SCR_SIZE_256x256,
+    GX_BG_COLORMODE_16,
+    GX_BG_SCRBASE_0xe800,
+    GX_BG_CHARBASE_0x00000,
+    GX_BG_EXTPLTT_01,
+    3,
+    GX_BG_AREAOVER_XLU,
+    0,
+    FALSE
+};
 
-static const BgTemplate sBgTemplate2 = {0, 0, GF_BG_BUF_SIZE_256x256_4BPP, 0x0000, GF_BG_SCR_SIZE_256x256, GX_BG_COLORMODE_16, 30, 0, 0, 2, 0, 0, 0};
+static const BgTemplate sBgTemplate2 = {
+    0, 0,
+    GF_BG_BUF_SIZE_256x256_4BPP,
+    0x000,
+    GF_BG_SCR_SIZE_256x256,
+    GX_BG_COLORMODE_16,
+    GX_BG_SCRBASE_0xf000,
+    GX_BG_CHARBASE_0x00000,
+    GX_BG_EXTPLTT_01,
+    2,
+    GX_BG_AREAOVER_XLU,
+    0,
+    FALSE
+};
 
-static const BgTemplate sBgTemplate3 = {0, 0, GF_BG_BUF_SIZE_256x256_4BPP, 0x0000, GF_BG_SCR_SIZE_256x256, GX_BG_COLORMODE_16, 29, 0, 0, 3, 0, 0, 0};
+static const BgTemplate sBgTemplate3 = {
+    0, 0,
+    GF_BG_BUF_SIZE_256x256_4BPP,
+    0x000,
+    GF_BG_SCR_SIZE_256x256,
+    GX_BG_COLORMODE_16,
+    GX_BG_SCRBASE_0xe800,
+    GX_BG_CHARBASE_0x00000,
+    GX_BG_EXTPLTT_01,
+    3,
+    GX_BG_AREAOVER_XLU,
+    0,
+    FALSE
+};
 
-static const BgTemplate sBgTemplate4 = {0, 0, GF_BG_BUF_SIZE_256x256_4BPP, 0x0000, GF_BG_SCR_SIZE_256x256, GX_BG_COLORMODE_16, 31, 4, 0, 0, 0, 0, 0};
+static const BgTemplate sBgTemplate4 = {
+    0, 0,
+    GF_BG_BUF_SIZE_256x256_4BPP,
+    0x000,
+    GF_BG_SCR_SIZE_256x256,
+    GX_BG_COLORMODE_16,
+    GX_BG_SCRBASE_0xf800,
+    GX_BG_CHARBASE_0x10000,
+    GX_BG_EXTPLTT_01,
+    0,
+    GX_BG_AREAOVER_XLU,
+    0,
+    FALSE
+};
 
 static const WindowTemplate sWindowTemplates[3] = {
     {2, 26, 21, 6,  3, 3, 0x3EE},
